backend/service/requirements.txt
pymysql
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pg8000==1.30.3

# Google Cloud
google-cloud-sql-connector==1.5.0
cloud-sql-python-connector[pg8000]==1.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Logging & Monitoring (optional)
python-json-logger==2.0.7

# Testing (optional for development)
pytest==7.4.3
httpx==0.25.2